% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{run_model}
\alias{run_model}
\title{Run a simulation of fast and slow movement.}
\usage{
run_model(
  scenario,
  popsize,
  nItems,
  landsize,
  nClusters,
  clusterSpread,
  regen_time,
  tmax,
  genmax,
  paramBallisticGammaA,
  paramBallisticGammaB,
  paramBallisticKappa,
  paramSearchGammaA,
  paramSearchGammaB,
  paramSearchKappa,
  range_perception,
  costMove,
  tSearch,
  pSearchSlow,
  pSearchFast,
  pStrategy,
  nThreads,
  dispersal,
  mProb,
  mSize
)
}
\arguments{
\item{scenario}{The scenario: 0 for ecological scenario only, or
1 for evolved probability of switching to area restricted search.}

\item{popsize}{The population size.}

\item{nItems}{How many food items on the landscape.}

\item{landsize}{The size of the landscape as a numeric (double).}

\item{nClusters}{Number of clusters around which food is generated.}

\item{clusterSpread}{How dispersed food is around the cluster centre.}

\item{regen_time}{The item regeneration time.}

\item{tmax}{The number of timesteps per generation.}

\item{genmax}{The maximum number of generations per simulation.}

\item{paramBallisticGammaA}{Alpha parameter for a step length distribution
used to draw steps for agent ballistic movement.}

\item{paramBallisticGammaB}{Beta parameter for a step length distribution
used to draw steps for agent ballistic movement.}

\item{paramBallisticKappa}{Concentration parameter of a von Mises distribution from
which turning angles are drawn in radians, for ballistic movement.
Should be smaller than \code{searchAngle}.}

\item{paramSearchGammaA}{Alpha parameter for a step length distribution
used to draw steps for agent searching movement.}

\item{paramSearchGammaB}{Beta parameter for a step length distribution
used to draw steps for agent searching movement.}

\item{paramSearchKappa}{Concentration parameter of a von Mises distribution from
which turning angles are drawn in radians, for searching movement.
Should be greater than \code{ballisticAngle}.}

\item{range_perception}{The range at which agents detect items.}

\item{costMove}{The energetic cost per distance moved.}

\item{tSearch}{The duration of area restricted search; this is fixed, while
the probability of switching to this mode varies.}

\item{pSearchSlow}{The probability of switching to search mode for so-called slow
agents.}

\item{pSearchFast}{The probability of switching to search mode for so-called fast
agents.}

\item{pStrategy}{The initial proportion of fast individuals in the population.}

\item{nThreads}{How many threads to parallelise over. Set to 1 to run on
the HPC Peregrine cluster.}

\item{dispersal}{A float value; the standard deviation of a normal
distribution centred on zero, which determines how far away from its parent
each individual is initialised. The standard value is 5 percent of the
landscape size (\code{landsize}), and represents local dispersal.
Setting this to 10 percent is already almost equivalent to global dispersal.}

\item{mProb}{The probability of mutation. The suggested value is 0.01.
While high, this may be more appropriate for a small population; change this
value and \code{popsize} to test the simulation's sensitivity to these values.}

\item{mSize}{Controls the mutational step size, and represents the scale
parameter of a Cauchy distribution.}
}
\value{
An S4 class, \code{simulation_output}, with simulation outcomes.
}
\description{
Run a simulation in which agents move on a landscape with
discrete food items and switch to area restricted search with some
probability, after encountering a food item.
}
